fix: adjust composition parts, move oxidation checks to utils
