+++ matminer/featurizers/composition/__init__.py
@@ -1,10 +1,12 @@
 # Core Featurizers
+# --> Element type features
 from matminer.featurizers.composition.element import (
     ElementProperty,
     Stoichiometry,
     ElementFraction,
     TMetalFraction,
 )
+# --> Ion type features
 from matminer.featurizers.composition.oxidation import (
     IonProperty,
     CationProperty,
@@ -13,24 +15,33 @@ from matminer.featurizers.composition.oxidation import (
     ElectronAffinity,
     is_ionic,  # needed for a test
 )
+# --> Orbital idea based features
 from matminer.featurizers.composition.orbitals import (
     BandCenter,
     AtomicOrbitals,
     ValenceOrbital,
 )
-from matminer.featurizers.composition.cohesive import (
+# --> Thermodynamics based features
+from matminer.featurizers.composition.thermo import (
     CohesiveEnergy,
     CohesiveEnergyMP,
+    YangSolidSolution,
+    Miedema,
 )
+# --> Packing based features
 from matminer.featurizers.composition.packing import (
     AtomicPackingEfficiency,
 )
+# --> Composite features built from other parts
+from matminer.featurizers.composition.composite import (
+    Meredig,
+)
 
 # External Featurizers
-from matminer.featurizers.composition.miedema import Miedema
-from matminer.featurizers.composition.meredig import Meredig
-from matminer.featurizers.composition.yang import YangSolidSolution
-from matminer.featurizers.composition.roost import RoostFeaturizer
+# --> Roost based features (https://arxiv.org/abs/1910.00617)
+from matminer.featurizers.composition.external.roost import (
+    RoostFeaturizer,
+)
 
 __author__ = ("Logan Ward, Jiming Chen, Ashwin Aggarwal, Kiran Mathew, "
               "Saurabh Bajaj, Qi Wang, Maxwell Dylla, Anubhav Jain")
