+++ matminer/featurizers/composition/composite.py
