+++ matminer/featurizers/composition/thermo.py
@@ -6,14 +6,298 @@ from functools import reduce
 
 import numpy as np
 import pandas as pd
+from pymatgen import MPRester
 from pymatgen import Element
 from pymatgen.core.composition import Composition
 from pymatgen.core.periodic_table import get_el_sp
 
 from matminer.featurizers.base import BaseFeaturizer
+from matminer.featurizers.utils.stats import PropertyStats
+from matminer.utils.data import (
+    MagpieData,
+    MixingEnthalpy,
+    CohesiveEnergyData,
+)
 
 module_dir = os.path.dirname(os.path.abspath(__file__))
 data_dir = os.path.join(module_dir, "../..", "utils", "data_files")
+
+
+class CohesiveEnergy(BaseFeaturizer):
+    """
+    Cohesive energy per atom using elemental cohesive energies and
+    formation energy.
+
+    Get cohesive energy per atom of a compound by adding known
+    elemental cohesive energies from the formation energy of the
+    compound.
+
+    Parameters:
+        mapi_key (str): Materials API key for looking up formation energy
+            by composition alone (if you don't set the formation energy
+            yourself).
+    """
+
+    def __init__(self, mapi_key=None):
+        self.mapi_key = mapi_key
+
+    def featurize(self, comp, formation_energy_per_atom=None):
+        """
+
+        Args:
+            comp: (str) compound composition, eg: "NaCl"
+            formation_energy_per_atom: (float) the formation energy per atom of
+                your compound. If not set, will look up the most stable
+                formation energy from the Materials Project database.
+        """
+        comp = comp.reduced_composition
+        el_amt_dict = comp.get_el_amt_dict()
+
+        formation_energy_per_atom = formation_energy_per_atom or None
+
+        if not formation_energy_per_atom:
+            # Get formation energy of most stable structure from MP
+            struct_lst = MPRester(self.mapi_key).get_data(comp.reduced_formula)
+            if len(struct_lst) > 0:
+                most_stable_entry = sorted(struct_lst, key=lambda e: e['energy_per_atom'])[0]
+                formation_energy_per_atom = most_stable_entry['formation_energy_per_atom']
+            else:
+                raise ValueError(f"No structure found in MP for {comp}")
+
+        # Subtract elemental cohesive energies from formation energy
+        cohesive_energy = -formation_energy_per_atom * comp.num_atoms
+        for el in el_amt_dict:
+            cohesive_energy += el_amt_dict[el] * \
+                               CohesiveEnergyData().get_elemental_property(el)
+
+        cohesive_energy_per_atom = cohesive_energy / comp.num_atoms
+
+        return [cohesive_energy_per_atom]
+
+    def feature_labels(self):
+        return ["cohesive energy"]
+
+    def implementors(self):
+        return ["Saurabh Bajaj", "Anubhav Jain"]
+
+    def citations(self):
+        # Cohesive energy values for the elements are taken from the
+        # Knowledgedoor web site, which obtained those values from Kittel.
+        # We include both citations.
+        return [
+            "@misc{, title = {{Knowledgedoor Cohesive energy handbook}}, "
+            "url = {http://www.knowledgedoor.com/2/elements{\_}handbook/cohesive{\_}energy.html}}",
+            "@book{Kittel, author = {Kittel, C}, isbn = {978-0-471-41526-8}, "
+            "publisher = {Wiley}, title = {{Introduction to Solid State "
+            "Physics, 8th Edition}}, year = {2005}}"]
+
+
+class CohesiveEnergyMP(BaseFeaturizer):
+    """
+    Cohesive energy per atom lookup using Materials Project
+
+    Parameters:
+        mapi_key (str): Materials API key for looking up cohesive energy
+            by composition alone.
+    """
+
+    def __init__(self, mapi_key=None):
+        self.mapi_key = mapi_key
+
+    def featurize(self, comp):
+        """
+
+        Args:
+            comp: (str) compound composition, eg: "NaCl"
+        """
+
+        # Get formation energy of most stable structure from MP
+        with MPRester(self.mapi_key) as mpr:
+            struct_lst = mpr.get_data(comp.reduced_formula)
+            if len(struct_lst) > 0:
+                most_stable_entry = sorted(struct_lst, key=lambda e: e['energy_per_atom'])[0]
+                try:
+                    return [mpr.get_cohesive_energy(most_stable_entry["material_id"], per_atom=True)]
+                except:
+                    # NOTE comprhys: f-strings do not work with pandas hence use of `.format()`.
+                    raise ValueError(
+                        "No cohesive energy can be determined for material_id: {}".format(most_stable_entry["material_id"])
+                    )
+            else:
+                raise ValueError(f"No structure found in MP for {comp}")
+
+    def feature_labels(self):
+        return ["cohesive energy (MP)"]
+
+    def implementors(self):
+        return ["Anubhav Jain"]
+
+    def citations(self):
+        return [
+            "@article{doi:10.1063/1.4812323, author = {Jain,Anubhav and Ong,"
+            "Shyue Ping  and Hautier,Geoffroy and Chen,Wei and Richards, "
+            "William Davidson  and Dacek,Stephen and Cholia,Shreyas "
+            "and Gunter,Dan  and Skinner,David and Ceder,Gerbrand "
+            "and Persson,Kristin A. }, title = {Commentary: The Materials "
+            "Project: A materials genome approach to accelerating materials "
+            "innovation}, journal = {APL Materials}, volume = {1}, number = "
+            "{1}, pages = {011002}, year = {2013}, doi = {10.1063/1.4812323}, "
+            "URL = {https://doi.org/10.1063/1.4812323}, "
+            "eprint = {https://doi.org/10.1063/1.4812323}}",
+            "@article{Ong2015, author = {Ong, Shyue Ping and Cholia, "
+            "Shreyas and Jain, Anubhav and Brafman, Miriam and Gunter, Dan "
+            "and Ceder, Gerbrand and Persson, Kristin a.}, doi = "
+            "{10.1016/j.commatsci.2014.10.037}, issn = {09270256}, "
+            "journal = {Computational Materials Science}, month = {feb}, "
+            "pages = {209--215}, publisher = {Elsevier B.V.}, title = "
+            "{{The Materials Application Programming Interface (API): A simple, "
+            "flexible and efficient API for materials data based on "
+            "REpresentational State Transfer (REST) principles}}, "
+            "url = {http://linkinghub.elsevier.com/retrieve/pii/S0927025614007113}, "
+            "volume = {97}, year = {2015} } "]
+
+
+class YangSolidSolution(BaseFeaturizer):
+    """
+    Mixing thermochemistry and size mismatch terms of Yang and Zhang (2012)
+
+    This featurizer returns two different features developed by
+    .. Yang and Zhang `https://linkinghub.elsevier.com/retrieve/pii/S0254058411009357`
+    to predict whether metal alloys will form metallic glasses,
+    crystalline solid solutions, or intermetallics.
+    The first, Omega, is related to the balance between the mixing entropy and
+    mixing enthalpy of the liquid phase. The second, delta, is related to the
+    atomic size mismatch between the different elements of the material.
+
+    Features
+        Yang omega - Mixing thermochemistry feature, Omega
+        Yang delta - Atomic size mismatch term
+
+    References:
+        .. Yang and Zhang (2012) `https://linkinghub.elsevier.com/retrieve/pii/S0254058411009357`.
+    """
+
+    def __init__(self):
+        # Load in the mixing enthalpy data
+        #  Creates a lookup table of the liquid mixing enthalpies
+        self.dhf_mix = MixingEnthalpy()
+
+        # Load in a table of elemental properties
+        self.elem_data = MagpieData()
+
+    def precheck(self, c: Composition) -> bool:
+        """
+        Precheck a single entry. YangSolidSolution does not work for compositons
+        containing any binary elment combinations for which the model has no
+        parameters. We can nearly equivalently approximate this by checking
+        against the unary element list.
+
+        To precheck an entire dataframe (qnd automatically gather
+        the fraction of structures that will pass the precheck), please use
+        precheck_dataframe.
+
+        Args:
+            c (pymatgen.Composition): The composition to precheck.
+
+        Returns:
+            (bool): If True, s passed the precheck; otherwise, it failed.
+        """
+        return all([e in self.dhf_mix.valid_element_list
+                    for e in c.element_composition.elements])
+
+    def featurize(self, comp):
+        return [self.compute_omega(comp), self.compute_delta(comp)]
+
+    def compute_omega(self, comp):
+        """Compute Yang's mixing thermodynamics descriptor
+
+        :math:`\\frac{T_m \Delta S_{mix}}{ |  \Delta H_{mix} | }`
+
+        Where :math:`T_m` is average melting temperature,
+        :math:`\Delta S_{mix}` is the ideal mixing entropy,
+        and :math:`\Delta H_{mix}` is the average mixing enthalpies
+        of all pairs of elements in the alloy
+
+        Args:
+            comp (Composition) - Composition to featurizer
+        Returns:
+            (float) Omega
+        """
+
+        # Special case: Elemental compound (entropy == 0 -> Omega == 1)
+        if len(comp) == 1:
+            return 0
+
+        # Get the element names and fractions
+        elements, fractions = zip(*comp.element_composition.fractional_composition.items())
+
+        # Get the mean melting temperature
+        mean_Tm = PropertyStats.mean(
+            self.elem_data.get_elemental_properties(elements, "MeltingT"),
+            fractions
+        )
+
+        # Get the mixing entropy
+        entropy = np.dot(fractions, np.log(fractions)) * 8.314 / 1000
+
+        # Get the mixing enthalpy
+        enthalpy = 0
+        for i, (e1, f1) in enumerate(zip(elements, fractions)):
+            for e2, f2 in zip(elements[:i], fractions):
+                enthalpy += f1 * f2 * self.dhf_mix.get_mixing_enthalpy(e1, e2)
+        enthalpy *= 4
+
+        # Make sure the enthalpy is nonzero
+        #  The limit as dH->0 of omega is +\inf. A very small positive dH will approximate
+        #  this limit without causing issues with infinite features
+        enthalpy = max(1e-6, abs(enthalpy))
+
+        return abs(mean_Tm * entropy / enthalpy)
+
+    def compute_delta(self, comp):
+        """Compute Yang's delta parameter
+
+        :math:`\sqrt{\sum^n_{i=1} c_i \left( 1 - \\frac{r_i}{\\bar{r}} \\right)^2 }`
+
+        where :math:`c_i` and :math:`r_i` are the fraction and radius of
+        element :math:`i`, and :math:`\\bar{r}` is the fraction-weighted
+        average of the radii. We use the radii compiled by
+        .. Miracle et al. `https://www.tandfonline.com/doi/ref/10.1179/095066010X12646898728200?scroll=top`.
+
+        Args:
+            comp (Composition) - Composition to assess
+        Returns:
+            (float) delta
+
+        """
+
+        elements, fractions = zip(*comp.element_composition.items())
+
+        # Get the radii of elements
+        radii = self.elem_data.get_elemental_properties(elements,
+                                                        "MiracleRadius")
+        mean_r = PropertyStats.mean(radii, fractions)
+
+        # Compute the mean (1 - r/\\bar{r})^2
+        r_dev = np.power(1.0 - np.divide(radii, mean_r), 2)
+        return np.sqrt(PropertyStats.mean(r_dev, fractions))
+
+    def feature_labels(self):
+        return ['Yang omega', 'Yang delta']
+
+    def citations(self):
+        return ["@article{Yang2012,"
+                "author = {Yang, X. and Zhang, Y.},"
+                "doi = {10.1016/j.matchemphys.2011.11.021},"
+                "journal = {Materials Chemistry and Physics},"
+                "number = {2-3},"
+                "pages = {233--238},"
+                "title = {{Prediction of high-entropy stabilized solid-solution in multi-component alloys}},"
+                "url = {http://dx.doi.org/10.1016/j.matchemphys.2011.11.021},"
+                "volume = {132},year = {2012}}"]
+
+    def implementors(self):
+        return ['Logan Ward']
 
 
 class Miedema(BaseFeaturizer):
